Update Test binstub to use new Test Suite
Update the Test Suite to use a format used within the SQLite package
present within the Smalltalk Git Mirror project files, in order to
better format this project tests.

Then update to Test Binstub to use this new file and instead of running
an invalid `gst-sunit` environment, run those commands directly, once
using it as a file and second as a collection of tests in order to get
more valid results.

== Notes:
- Update Package information to include files that were omitted and
  reference the Test Suite directly

- It is unknown what is the correct way to run these tests, since
  GNU-Smalltalk works differently to other Smalltalk IDEs and generally
  Smalltalk is not usually built this way

== References:
- [smalltalk/SQLiteTests\.st at master · gnu\-smalltalk/smalltalk]
  (https://github.com/gnu-smalltalk/smalltalk/blob/master/packages/dbd-sqlite/SQLiteTests.st#L238-L264)