Update README with Robot Simulator specification
Update README with the Robot Simulator specification that the project is
attempting to implement and contains several examples of how it works.
Add those examples as simple text files for Users that later download
this project can then use against the `bin/app` entrypoint in order to
evaluate the application themselves.