Add Robot CLI `run` and `processInput` methods
Add a ToyRobot Command Line Interface (CLI) with methods `run` and
`processInput`. The Run method determines the input stream to be used
based on whether or not a filename string was passed in as a command
line argument, otherwise it defaults to accepting standard user input.

The other method Process Input accepts the input source stream and a
robot instance to being processing that input, executing commands on
the Robot instance (with or without additional arguments). It loops on
this process, getting line by line from either User command prompt or a
read file and only quits when the User types `EXIT` or the end of file
is reached.

== References:
- [Read a data from an input Stream line by line in Smalltalk \- Stack
  Overflow]
  (https://stackoverflow.com/questions/42588982/read-a-data-from-an-input-stream-line-by-line-in-smalltalk)

- [Chapter 3: Smalltalk Statements]
  (http://esug.org/data/Old/ibm/tutorial/CHAP3.HTML)

- [How can i get a new line in smalltalk? \- Stack Overflow]
  (https://stackoverflow.com/questions/13932039/how-can-i-get-a-new-line-in-smalltalk)

- [tokenize \- Smalltalk, newline character \- Stack Overflow]
  (https://stackoverflow.com/questions/1598054/smalltalk-newline-character)